power management updates, build fixes when using PM
When CONFIG_PM_ENABLE is set, there was a syntax error due to using a
variable that had not yet been defined. Variable is now defined before
all usage sites.

Also reduced default CPU frequency to, maybe, reduce
power usage while not sleeping. This is guesswork, and could be wrong,
as sleepy systems tend to want to race through all available work while
awake so they can get back to being asleep, but guessing that a fair bit
of the work being done while awake is I/O-bound rather than CPU-bound,
so running at a lower clock might save _something_.

Made light sleep mode configurable, along with the delay between
readings. Light sleep mode, when enabled, causes serial debug to stop
working, which is inconvenient when debugging code, but beneficial in
production.

As a result of the longer default time between readings, switched the
delay to happen after the reading instead of before so that the
time-to-first-reading is fast.